Revamped the view which lists the episodes of a selected season. The info row was moved to the top above the title. I simplified the title and subtitle messages below each episode card to just Episode N. I think the season number and the episode number were repeated too many times at various locations of the UI. To make the info row size adjustable per view, I added a variable sizeScale to InfoLayoutHelper.addInfoRow() , which gets propagated to all downstream InfoRowItems. It also scales the layout_height of the LinearLayout InfoRow so all the tags (codec, rating, etc.) are properly sized. The reason for this change was I wanted to make the info row bigger on the episode list view but not in other views.
Just a minute detail, I changed the preferred ImageType from Poster to Thumb in LibraryPreferences. By doing so, I only need to change it once for Movies and Series but I would have to go into each of my other collections to change it from Poster to Thumb if I did not make that change.